Changes to custom Reports to support outputting to Excel/Word etc.
This is the first part of the fix in which we're moving the old vb6 activex client upto the server and convert over to .NET (previous check-in). I've then had to substantially rework how the custom reports run . 

A small history lesson. WE used to build up strings put them into input controls stream them to the client and then there was some javascript that would read these input in and put into a table (or an activex grid if you go further back). These values would also get read into a hidden grid which was then traversed to take the values and put into the client activex ready for output. This was very cumbersome.

In order to get this all running on the server I've modified things to we now populate a datatable and bind this directly to a gridview. This datatable is now also passed to the download_ouput in the homecontroller where the data is sent to the converted cleint side code.

At present this code still relies on having Office installed on the server, but there will be another imminent check-in where these objects are replaced so don't woryy too much at this stage about testing the actual ouput.

I've also changed some of the multi-dimensional arrays to be strongly typed collections as understanding that part of the code was an absolute mare.

2 known bugs which I'm going to raise separatly 1) the group with next does not put linespaces in and secondly the columns need some kind of formatting to the right size.