Added a jump to current month for calendar report as well as looking at this setting that is read from the defintion.
Whilst I was here I also did a bit of tidying up and got rid of a load of functions that are now uneccesary following the move of data generation to the server side. After I did think I saw that there was only one function requierd (ExportData) remaining so I moved it into the data.ascx file and zapped the budle for calendar report bundle.

Also removed the TimeRangeSelectedHandling from the daypilot control as this was causing a hover button for days in the schedule with no data.

Got rid of some commented out code and unused "blah" objects.