By default when you add a calculation to a report it sets it as size 0 and so was not getting displayed.
In Data Manager it always truncates numeric values so that a column with a width of 4 and a value of 16786 would get shortened to 1678. Discusssed with Mark and this is not what it should be doing so it now displays #### if the value is too large. (like Excel). Does mean a difference between DM and DMI but hey ho, we've got to make progress somehow.