A bit of formatting change (some enforced by the editor) and removed a couple of unnessary <% %> compile directives
Also added 2 .FindAll enumerators on the collection to avoid unneccessary loops around the collection. Also removed 2 if clauses on top of this as they are no longer needed (trying to find out which endifs needing zapping was a pain in the derrier because nothing is lined up correctly and as soon as you attemp to line things up the editor alters it on your behalf. Arrrggghh)