Slightly better string handling (using the & is processor intensive - think of it as a copy object for each one). I've only done a string format for the most heavily used because we're off to beta soon and I don't want to take too higher risk as you know who'd moan. This is a quick and easy performance boost we can finish later.
Also removed the clsGeneral class that was only used in two places (called the gettable function directly). On the face of it it's only one little class - how much performance can you gain from that - well its more a matter of having this class lying around pervents the garbage collector from doing its work properly so as the day goes on we'll have a biger and bigger memory leak. There are loads of other places where this should also be employed - we should sort them out as we find them.