Renamed and simplifed tablepriviledges and columnpriviledges classes. It may not look like a great deal, but those private properties were getting called literally thousands of times during load up time. This is a cleaner and more .NET way of doing things.
The rename is part of an ongoing personal campaign of mine to start getting some kind of standards into our naming conventions (and to finally release the shackles of VB6)

I've also noticed that part of this check-in has altered some whitespace items. This is not a deliberate attempt to make extra work in the code review. Sorry.