JIRA-3051 - When running a report with multiple columns that rely on udfs (calculations with child table->specific lines referenced) then the holding array for UDF was not being set correctly.
This is still not an ideal fix because it leaves the array starting with a nothing object which is a bit messy and a timebomb if the expression code is ever tidied up. Bit tricky to change quickly though as the whole array declaration is shared with the array that holds the runtime code for the expressions, which I don't really want to starting hacking around with right now. And franky neither do you!