Changed readonlyrecords to be a firehose cursor. A history lesson (which is more or less commented in the getrecords function above the change) is that older versions of the ADO connection object would only allow one firehose cursor per connection, so if you had a function within a function both of which opened recordsets then you'd get a big fat juicy error. This is no longer the case as we can define the connection to have something called MARS (multiple active resultsets (or something like that (Google would confirm either way))) which gets around this by managing connections as necessary).
The one catch with a forwardonly that it does not return a .recordcount because it won't know until the whole recordset it processed. This is not an efficient way of looking at data anyway, but hey guess what we do, which is why I've altered the one or two places that did use it to access the old slow way. (or swapped out to use a .bof .eof if all that's needed is a check to see if there are records)

Swapped the two functions in clsGeneral to be friend so we can start tightening up on the scope of some of the functions, which will help when it come to cleaning up the code.