Its our old friend of byref and byvals. (For those with a short memory this is due to a problem during the upgrade from VB6 where everything used to be passed as an object type and passed parameters could be altered. In .NET if you pass a recordset field value it needs to be passed by value otherwise teh casting operation fails.
Also set the preview on screen to always return false as this is no longer a valid option because we are downloading teh completed document to the client. (also had an intersting side effect of trying to display on the server. Whoops)

Also a bit of code tidying up swapping things over to RETURN instead of setting variables.