An error with session variable not set when accessing child find windows directly from the dashboard (i.e. not going into any recordedit beforehand)
Wrapped up in a nullsafestring. 

Having trawled though the code I'll be honest I can't see what the point of this lineage session variable is doing as it seems to be basically a dup[licate of teh cancelcoursedatacolumn session. I don't really want to hack all this stuff out at this stage, but this might be an idea to come back to if/when we clean up the training booking module.