JIRA-3320 - The recordedit submit was getting called twice when adding a new record.
There was a navigation property that was only ever being set and never used in any shape or form. If it was set to zero it caused a new record to resubmit and during that process lost the session("action") variable.

Since its not doing anything I've removed the whole kit and kaboodle because as it did was confuse things. Two areas that looked for the existance of the frmWorkAreaInfo form have been swapped to frmMenuInfo so the menu loads as it did before.