Enabled the backpage functionality. I have to say though that I'm not sure that we've got this back/parent/cancel/close functionality implemented in the most user friendly way.
As it stands now the back button here returns to record edit. It differs from the back button in SSI which will sometimes return to recordedit ajnd sometimes return to the dashboard. I bet you a jam tart we have to review this after the user have got their hands on it... :-)