JIRA-3621 - A long, long time ago (pre 3.5) we used to store OLE headers with a length of 300 bytes. This was then extended to 400 to handle longer file and directory paths. As part of a system manager save it attempts to upgrade these structures and then mark the db with the current version (<<V0002>>) so that upgrades don't re-run. Now I'm not quite sure how, but the QA database mentioned in this fault has not been upgraded, but the system manager has marked the db as v2, and hence it goes a pete tong when accessing the records.
Its not really a solution as I don't want top start putting update code in the UI, but I've put in some tighter error handlngt so that it no longer blows up. It just doesn't display structures that are the wrong format.

Also as part of this had to replace an on error with a try...catch and cast the ole object as a strongly typed object

There are still outstanding issues with OLEs mentioned in the associated work item. These will be addressed in following check-ins. This fix relates purely to JIRA-3621