Yuck - what a horrid one this was to track down. Firstly because of a loose apostraphee in the link name the broswer just blew up with no indication of where the error was.
JIRA-3657 - The second part was that the multi axis chart was blowing up because the seriesname was set to the wrong column (3 instead of 2) and also put protection in the bit that gets the table and column names as they can be set to 0.

While I was here I tidied up classes and renames .BaseTable to .BaseTableID for consistency.

All these problems were noted when testing the Kennedys system before going to pilot.