This is the groundwork for moving some of the classic ado code over to ado.net. The sysprotects stuff must bite the dust in order to use ado.net as the cmoving this onnection is openly opened when required and automatically closed by the windows framework (garbage disposal) and hence the temp tables are zapped every time.
This will be subsequent check-ins removing #sysprotects from all the stored procedures, but I wanted to seperate out this bit of code because it was done in the fat clients.

Now after the system anad security manager saves complete a stored proc is called which reads all the sysprotect settings and copies them into a temporary cache table. On the face of this this shouldn't have too much of a performance gain, but it does as teh syprotects is actually a very complex veiw which cannot be indexed (and also is a composite of all other databases on the server)

Eventually I'd like to move away from using sysprotects entirely as its a bit of a resource hog - we should be looking at using sys.database_permissions and sys.server_permissions (cleaner, quicker and easier to understand)

To test/ensure this is working all you can do is run the update script and save system manager as the rest of the code that will use this has not yet been completed.

P.S. spASRGetMetatadata is included as it is part of the ado removal and the project file was checked out to include the other files.