Finishing the groundwork for moving code over to ADO.NET. The stored procedures have now been moved over to use the cached copy of asrsyprotects instead of the #sysprotects. I've run the scripts on our dev databases but you'll need to run the latest script and save system manager on any other databases you wish to use.
Basically now all the references to sysprotects now have any extra where clause to the UID. In most functions this uid was already calculated but in one or two I've had to add this code.

Building the update script was pain in the rear. We really need to sort out our deployment stuff.

The next chck-in will see us switch a lot of this code over to ADO.NET. Exciting eh?