This is the beginning of moving the old classic ado connection to ado.net. Firstly is the modifications in the clsDataAccess class to now have a secondary connections which is ado.net. We still need to support the old ado for the moment as there's too much code to convert in one hit, ut this can be replaced over time (or if you get bored...)
Thje majority of the converted code has swapped classic recordsets to .net datatables as these are the closest matches in terms of functionality - notably different in that instead of using .movenext these are enumable pieces of data so we swap to a foreach.

I've swapped out to a dataset for the populateMetada function because we can retrive more than one bit of data at a time (saving roundtrips to the server). This can and should be employed in more areas and I will look at using more of these over the coming months.

Now that we are using SQL2008 and above we can now define user defined types in SQL. This allows us to pass datatables back and forth from the server. And its been used in the setupTableCollection because the old command string could not be passed as a parameter and we needed it convertinbg to a sp because its doing more than just a simple select nowadays. I think we should be looking at using more of these in the future as its a nice way of not having sql to constantly re-read the same data (i.e. all the permissions gubbins we still have lying around)

Defsel has been converted over to a datatable, however I haven't swapped the actual grid to bind to it directly. This would seem to be a relatively simple thing (and would result in a performance boost) to do in the near future - lets get rid of that tabletoGrid function. I just don't have the time or the desire to faff around with styling at the moment. Volunteers?

Other minor changes in this changeset
1) Use of uppercasing in the stored procs so it doesn't have to be done in each row down in the server dll.
2) Removal of unneccessary ADODB. This will make the final removal a bit simpler to track when this is finally removed.

Things to note
1) There's now an autologin if you open a second tab. Might be useful. The logoff is still a bit unfinished because when you log out you could still bu***er up other windows that may be open.

For reviewer
There should be no changes to security, but if there are any concersn this is where they'll be.
I would hope the areas where it been converted run a tad faster, even though this will probably only be noticeable when it goes under load.
You'll need to run the update scripts on any new database you're using.