History - Some calculations require user defined functions to be created on the fly so that they can work (sepcific line functionality is the only one still using it). To do this a pass of the expression code generation was done (including reading the definition from the db) and then a second pass (also re-reading) to do almost exactly the same thing but wrap up the code in a udf.
There was no reason for this to be done twice - instead a udfarray is passed into the runtime code and both operations can be done together.

To check validity of this run some reports/mail merges and build expressions with specific line functionality.

P.S. Got rid of setting of a unnecessary (and broken) connection when testing an expression