Moved the logoff from the home to the account controller because while it was in home it forced a frefresh of site.master and in turn this pulled up the menu.ascx which still has some of teh old timeout problems on the classic ado connection.
Also the logoff was not recording the action in the database, so put this into a new ado.net routine. (the session_end still uses the old way)

There still needs a bit of thinking about how we log out if tey have multiple windows open, do we prevent the actual logout or auto logout the other windows? For another day!.