Swapped the menu items out to a strongly typed collection, and rejigged so that it does not request data from the server every time menu.ascx refreshes itself (which was everytime you clicked a menu option!)
Zapped the get misc info as it returned values that were not used anywhere (I think this is an old legacy bit of code that was used when we used to implement our own security lockout functionality instead of using SQL inbuilt stuff)

Started to cast values when returning from recordsets so that we can look at setting option strict on (performance and code safety). Have not actually implemented yet as there are a few areas which don't quite work yet.

Zapped sp_asrintgetSystemPermission. Function not longer used and removalk should have been in a previous check-in but wasn't.