Removed classic ado from running of calendar reports.
Teweaked a bit of code that used .moveNext in the middle of looping through the recordset to run some code at start of loop and end. ADO.NET doesn't have a movenext concept.

Removed the recordsettodatatable calls.

Cleaned a few <FORM> <INPUT> to lowercase.

Also started with option explicit so moved a few shorts to integers.