Finally removed ado from the project file, so the server dll is now completely devoid of any of the old wubbish ado.
As part of this check-in there are changes which basically move functions into the general and base classes as modules do not support inheritance (and hence no dataaccess class)

Also as part of this option explicit has been turned on some files where it was relatively simple to do so. (Hence lots of cint( and .tostrings. Also a lot of short to integer types have been tightened up.

The diary.checkaccestosystemevents is no longer called as part of the class initialisation as we won't have connection settings at this point, so made it public and called manually.

The global variables gADOCon has been removed and well as a few connection properies in classes.