There were uneccessary calls to the database to get a table name. Swapped some calls from the dmi.net project to use the cache in the server dll.
In order to support this I've had to expose the GetTableName in the base class and change some of the dll classes to look at this one instead of teh other one in clsGeneral. There's a very good arguement that GetColumnName and getDataType should move to this location too, but for this check-in I'm leaving where they are as I can envisonage a bit more fiddling round with this when we try and get rid of these pesky global variables.

Droppedd the sp_ASRIntGetTableName as it is no longer used anywhere in the app. I've run the latest script on npg_openhr8, but if you're using your own db - well you know the drill.