Converted loading of user config to ado.net
This is a bit of hotch potch fix though. I've done the hard work by putting all the calls through a function in the dll, but we have two ways of getting settings at the moment 1) The way you see here which is to read all the settings sequentially. And 2) In the PopulateMetaData function read all at startup and access these - which is much faster and avoids load on the database but it isn't updated from when they first log in. I think we'll have to revisit this later when we start looking at how we read in the module parameters (training bokking being a particular pig)

Added an enumerator for record edit start types, but have now rolled the implementation everywhere because it would just take too long, and I'm tired, hungry and getting just a little bit cranky.