Converted record edit insert, update and delete over to use ado.net.
There was a lot of deadlocking code checks which I've now moved into the dataAccess class. It will retry five times and then throw an exception to the calling class.

This then means a lot of code has been zapped from the homeController.

Because we're using try...catch we can't have an on error sitting side by side (which was an really ignore all errors anyway). Once the ado.net conversion is complete everything should be wrapped in the new try catches.

As part of this I have noticed a bug (pretty sure it's been there for a few versions now) is that if one user is doing a big global update (or a while 1=1 loop in management studio in my case) itc creating lots of deadlocks even if its not directly affaecting the record we're updating. I think this may be some quirky stuff with the wqay our triggers and udfs are complied. May be worth investigating at a future date.