JIRA-3666 - Was crashing when running a particular report.
This report was somehow corrupt and the working pattern was somehow setup as a numeric column and hence was formatted as a numeric (which won't work when you put a string into it!!!) Added an extra isnumeric check just to be on the safe side even though in a normal system this should never be called.

Also as part of this I noticed that the unique code column was also being set to ######. This is because the column is a numeric(10) and they are formatting as a numeric(6). Better to show an out of range than incorrect data (this is the way Excel works)