First off a bit of house keeping with try catch on the execute and making a bit more ready for option strict with variables of correct type and byvals.
Meaty part of the fix is in CreateMutipleChildTempTable(). The loop was wrong so that it always looked at the top row of data - renamed the rather unhelpful i variable to iChildRowCount

Also instead of firing off each sql command seperately concatenate and fire in one go, avoid lots of server round trips. This could be made faster again by inserting just as values, but we'd need to rejig the logi because the amount of columns can vary depending on the data and SQL doesn't like that.