Finished (well started too) off output options for custom reports. We should now be able to output nicely to Excel from any custom report.
Added a base class for the reports and the report outputs ready for the other output types and reports that are coming in an imminent check-in.

Removed clsColumn and replaced with the defined column class so as to avoid duplication of code (and to make understanding things a tad easier.

We now generate the output file name based on a temp name and the only time we look at the user defined output name is when we stream it down to the client.

Added UtilityType and FormatType enums in a couple of places to tighten up our code.

Did a bit of housekeeping and zapped some unecessary/commented out code.