JIRA-3373 - Was not adding the summary rowsd for Bradford into the report. Basically needed to add the staf number as a sort field and then to alter the aryTotal and aryCount to display the calculated Bradford factors.
Renamed (and hence taken from teh base class the ColExprID to be the ID column so that the extensions class can do its magic with the getbyindex and getbyID extensions.

Also did a bit of code tidying up whilst I was at it. ByRef and ByVals and a bit more work on getting ready from option strict on.