JIRA-3380 - The standard reports have the ability to prompt for date range information before displaying the report definition. However after discussion this seems a rather pointless excercise since the dates are displayed on the defintion screen anyway.
You can define a prompted calculation which was more complex than just prompting for the date (e.g. if department = "Sales" then 'thisyear', else 'last year'. However this doesn't tie into the actual filter of teh data so could be somewhat misleading. This is a slight functional change from 5.2.

So anyway what it now does it calculate the default prompted value and populate the main definition screen with those values. The menu calls the def screen directly rather than via prompted values so it now behaves in a similar way to other defsel screen (hence I can remove the standardreport variables from promptedvalues.

Also as part of this I've added a prompteddatevalue enum, zapped the GetStandardReportDate function and moved the calulate date value type to a function in ASRFunctions so that standrd and custom reports use the same code.