JIRA-3715 - Here be a valuable lesson in the difference between ByRef and ByVal and why we need to be strict using them. The psJoinCode is altered in this function when having child and parent columns on the same report.
Yuck - As a general rule functions should not be altering state of passed in variables. Given time this function should be tweaked so it returns proper values or at least sets module level variables.

To test (or break in previous build) Create a report with a child table and a sort on that child table which references back to the parent. Example *PR And Absence on npg_openhr8