JIRA-3756, 3757 - Now displays the record description if it can. The find window will display "My Dashboard"
I'm not certain that we've got the design of this correct as it displays in different places depending on tiles, wireframe or winkit mode and make more sense in some than in others.

The behaviour of the back button is something to ponder about here, because if that works perfectly then maybe its should display "My Dashboard" but the existing previous caption as you ahven't actually changed the record context at this point. Probably needs part of a wider review.