JIRA-3749 (Possibly more, they just haven't found them yet).
Basically this is a problem with loading the stdrpt_def page from both within the workframe and the optionframe - the direct references to frmDefiniton and frmAbsence were then failing because there were two of them in the dom. Simply issuing the LoadEmptyOption thingymebob didn't clear out the cache so its necessary to call via $("#frm_xxxx or else it'll blow up.

Another bug sorted is that filter and picklist selection would error as the baseHidden was missing from frmRecordSelection (don't think QA have spotted that one yet)

Took the opportunitey to clear out a couple of unneccessary forms and functions