JIRA-3358 - Added functionality to highlight weekends in calendar reports. This follows on from task 4864 of taking a snapshot of the DayPilot control.
A class is added to dates identified as weekends which can then be toggled.

As part of this I've had to update the .net framework for this project to 4, which has meant the removal of some of the AspNetHostingPermission attributes (what they were there for I don't know - all they did was occasionally break)

Styling of the options to be undertaken after other options are added. (focus on the functionality at present)