JIRA - 3559 - Implement bank holidays into calendar reports. Works slightly differently to how it used to in 5.2 as it takes advantage of being able to display multiple events for a single person over one day. Each bank holiday is now displayed as an event type (or absence) and is added to the legend of events. As a result there is no need for the show bank holidays checkboxes.
I've also tweaked the display to show the whole absence colour instead of just a line as it was a bit difficlt to see when you had short absences (of a day). I still need to review the colours chosen (you can argeue they are garish) and how we can best render the captions.

I've had to alter the daypilot control to handle event types so as to differenciate between bank hols and "normal" events.

Working pattersn still to be implemented.