JIRA-3564 - Report boundaries ere being ignored.
I've had to completely swap out the monthpicker control because the old one just didn't cut the mustard in that you couldn't disable any of the months or limit the year range. This one is more functional but could do with someone attacking it with some crayons.

Removed the old monthpicker from the project