Now outputs the calendar report to Excel. I've had to amend the DayPilot control to expose a few of its internals so that we can enumerate through its data when hoofing out to Excel.
Moved the calendar binding event out to a separate class so that it can be shared by both the preview and the document export

I've had to add a reference to the ASPOSE cells in the dmi project. I hope this doesn't aggrevate the build locking issues we currently have.

Work still to do (so don't bother looing at. 1/2 days need cells extending - will currently crash :-(
Multiple events on the same day will currently overlap :-(
Email is not emailing.
These will come in a subsequent check-in.