Clean up some of the warnings that visual studio is giving us when compiling. (Note - as soon as I cleaned up some of these warnings new ones appeared :-(
Tokk the opportunity to put option strict on Tablre and ColumnPrividleges - this should yield a tiny performance, but how much I'm not sure.

Removed a few byval and byref references and some functions that were never used.

In class chart and multiaxis chart it looks like I've just changed a comment line - but I've done FAR more than that... The visual studio compiler sees a ''' as some kind of xml markup and therefore throws a paddy - simple fix - just remove one of the ' and voila.