JIRA-3672 - Dates being displayed with a timestamp.
Bit of a faff this one, just to do a bit of formatting, but we have to implement the IFieldMergingCallback in order to intercept the merge, we can then kick the values around to our hearts content. I've taken the liberty to pass the column definition of the merge through to the run class at this point because very soon we will need to start applying specific formatting options (otrher JIRAs have been raised)

This should also be region "safe" although I can't say I tested it properly.