Error generating the email address since the move over to ADO.NET as the parameter list to the stored proc was badly formed. (wrong name and wrong size)
Also added some try catching in the bit where it executes the merge and passes back to the user if necessary.

Whilst I was here I've tried removing the enum from util_run in an attempt to reduce that damn warning about differenent versions of the dll. May be something to do with referencing an enum directly before it compiled. Maybe, I dunno, we'll see.