Changed a bit of training booking over to ado.net. Actually the hardest bit of this was to find out how and where in the application the code got fired from, so if you're reviweing this I'll save you a lot of heartache and tell you...
You'll need to ensure that there is a course available on the db, so lets go create one (create an avialble course of Excel Skills). Now select a current employee and create a training need record for this skill. Once this is done the find window book a course button should become enabled. Click on it and your recently enetered course should be avialble from the available courses grid. Select one. It will then ask for booked or provisional, click on either. Now the action is set to "SELECTBOOKCOURSE_2" and this code will now fire.

Easy eh?

Also tidied up the update script a tinsy bit.