Removed the classic ado from the login and change password functions. Note: The formatting is inconsistent between forced and normal change at present and subsequent jiras/changesets will follow to tidy these up.
Items to note. Two stored procs removed - The minimum length is not relevant anymore as this policy is controlled by the inbuilt sql security engine. This was an overhang from when we had our own in SQL7 andSQL2000.
Similarly the intpasswordok is also redundant as this was used to track complexity when we also did our own in the above environment.
The login process checks will have a bit more work on them to tidy and streamline (after I've had a nice relaxing weekend)
Removed the get native client stuff and checks as they are no longer necessary with ado.net
sp_password is no longer used as the sqlconnection object gives us in-built change password functionality