Misc JIRAs - Essentially a rejig of the order of the login logic. Prior to this there was a whopping sp_asrintchecklogin proc that did the whole thing and was a tad unwieldly. Now split into a smaller proc which returns details about the current login and the web app decides what permissions they should be using (rather than the stored procs). The decisions in the accountcontroller login are now lumped into one area before it goes off and populates the metadata.
This has involved shifting round the permissions to the sessioncontext class and modifications to various classes to support this.

The spASRIntGetActualUserDetails has been simplified to call the fat client version (they both more or less did the same (well, I say that, the Intranet one didn't work for windows authentication)

Version comparisons now use the inbuilt version class, however I've had to overload it because of the inconsistent way that we store our version numbers in the db.

Introduced some input tags on the main forms to support dmisingle and dmiuser 

sp_lockcheck was not in the source control library so I've added it.

There will be subsequent check-in to this work and this is foundational - this does more than just fix up the windows authentication.