The login and out f the app now goes through a track user function - this replaces the old intcheckauditaccess malarkey.
No specific JIRA has been raised for this, but this is groundwork for a number of other ones.

The spASRGetCurrentUsersFromMaster used to be generated by the systme manager - this is now a static procedure - so if you're going to save system manager yuou need to make sure you've got the latest version.