JIRA-3828  - Copying records did not copy over the old columns. The problem arose with changes from upgrading from 5.2 as previously we used to have a client side activeX that read the streamed values down in the client and created a secondary update to the table. This has now been moved to be an upload and download functionality and hence we don't have these ole values available when saving a record.
So the fix (ish) is to save the new record without any ole (no changes) and then if its a copy copy the ole columns from the original record by calling a new postsave stored procedure.

We track the original record id and pass this through the data submit. This is then passed into the stored proc and a sql string built to update from the existing record and then executed.

Note: Permissions are not evaluated at this stage. If you can't update a sepcific column an error will be returned to the client. The initial copy record will be successful however. This is different to 5.2 which would have looked through the permissions and cut out any columnd that we don't have access to. This is feasible to implment, but may take some time and if necessary should be raised as a seperate task/fault.