JIRA-3950 - Problem running a specific calendar report. The JIRA fault states that its to do with having set to preview and data only. Do not listen to that. That's bunkem.
The actual problem was that the report in question had the end date of the report before the start date and hence no data was getting returned. The problem was that there wasn't any reporting back of the problem to the client. So I've bolted in the existing raisewarning function and removed a ton of unwanted crappy code that didn't work.