JIRA-3973 - If there was a problem finding or executing the stored procedure it was not being error handled properly (our good old friend of on errors and try catches getting confused.
On a side note (and I've put this into the JIRA and theerfore might get raised as a separate fault (or just failed) I've noticed that the server and database name are not being passed through to this function - instead they are coming straight from the config file - this may or may not be a big problem depending on whether we tell the customrs about the ability to goto a different server/db on the url line). This is because this code originally came from SSI where the db and server settings always came from a config file and could not be changed. Oh for such days again...