The system only now checks for messages wehen the user clicks something, i.e. there is no heartbeat anymore. Thus the system manager can save to his hearts content.
You will need to run a system manager save and run all the update script before this all works.

A couple of caviats here. 1) I don't like this. I don't like throwing in an extra database call on every click. This is clunky and liable to cause performance issues. We need to handle this properly in 8.1 with a better technology and push from the server.

2) I'm not sure how usable this will be in the real world because the send message can only be send to users that the system manager knows are logged in, which won't include the web ones. If there's a DM then all works ok.