Moved some of the core metadata collections into the session class as to avoid spilling over into other sessions on the server.
Inserted a baseclass for modulespecifics - this is a shorthand so I don't have to change too many function names, with a bit of a cleanup these module classes can be converted to structures with the read in a separate class.

Moved a lot of functions out of clsGeneral to BaseReport and BaseForDMI and altered the references accordingly. This is because I'm looking to get rid of clsGeneral and have all of its functions as an inheritable object. 

Had to turn strict off temporarily on the baseclass until I can sort out the getcolumnpriv class. Boo, hiss.

There should be no functional changes in this changeset. If you notice any differences then that's bad news :-(