Changed the way the smtpclient picks up the settings. They are now read from the smtp setup instead of app settings. It takes more parameters than the smtpclient in clsOutputRun because its the aspose variant rather than the base .NET one. (The effect of the settings is the same though)
In order to get this running there are a few things you need to do. Either get your IIS box setup to send whcih may involve quite a bit of dicking round on the network with permissions and firewalls anbd routing etc), or you need to install the role of SMTP Server on the IIS box and then configure stmp mail to go through the localhost. How this is done is different for every version of Windows - I'd advise hitting Google on how to do this.