JIRA-3552 - Problem with having OLE controls from a table other than the base table. Modified so that it now only read the values as part of the getintRecord (as opposed to read them twice via the getstreamproperties) so should be a little quicker performing. Converted the loading of oles from a loose object to a strongly typed collection.
It will now displays the foreign columns but makes them readonly (pictures display, but are unclickable, oles show the filename, but are also unclickable) - a clever little fudge inserted into the spgetscreencontrols2 procedure. You will need to run the latest script on any database you wish to try this on.