Now hopefully works as suggested by JP i.e. display "No Pending Workflows" instead of trying to get rid of it in non-ssimodes.
There could be a clever way that NPG could come up with to make this work as he wants it to - i.e. if there are No Pending Steps then dont show anything in DMI.

I couldnt get it to work but I wa going down the route of showing it briefly then fading out to blank screen and reset button bar.