JIRA-4011 - Problem copying records where you don't have access to the physical base table. Was caused by the post copy record functionality for the OLE columns. Resolved by ensuring that the stored proc is executing as a sufficiently privedlged used (i.e. the dbo)
The whole of copy record needs a bit of a rethink/clarification of requirements. After consultation there seems to be a bit of confusion what should and what actually is when you copy a record. If you only have access to half the record details (through security or through a view definition) then what bits should be copied. Also since child details are not copied, it often looks odd that your copied record is quite different. Should OLEs be copied? You probably wouldn't want a photo copied, but you might want terms and conidtions document. 

On a side note - before you say just do what it does in Data Manager then you need to be aware it runtime crashes! (
Create a embedded document on PersRecords.
Do NOT add to the current employees view, but DO paint this onto a screen that is used by current employees view. 
Ensure that your DM user has full access to the current employees view, but no access to personnel records, or any access of any kind to that column you've just created.
Go into DM and select current employees - notice your new column is visible on the screen, but is disabled. Now copy the record and save.